Revert "Cleanup styling code."
This reverts commit dd862c0608cab855ef8e74429439da6c9185c404.
